+++ resources/release-notes.md
@@ -1,3 +1,7 @@
 # Software
 
+- Improve handling of "croppable" data (inheriting from `mag.mixin.Croppable`) with utility methods `mustBeTimeFilter` and `convertToTimeSubscript`
+
+# GitHub
+
 - Add template for bug issues